feat: update project structure and directory
