+++ app/(landing)/layout.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function HomeLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <div>{children}</div>;
+}
