+++ components/landing-page/about.tsx
@@ -1,0 +1,46 @@
+import Section from "@/components/layout/Section";
+import { AboutImage } from "@/public/Images";
+// import { AboutImage } from "@/public/images";
+import Image from "next/image";
+
+export default function About() {
+  return (
+    <Section
+      id="about"
+      className="pb-12.5 pt-8.5 flex flex-col gap-y-4 md:gap-y-8 mt-14 mx-auto"
+    >
+      <h2 className="text-4xl md:text-5xl font-extrabold  text-center md:text-left">
+        About StreamFi
+      </h2>
+      <div className="flex flex-col h-fit md:flex-row items-start xl:items-center gap-6 md:gap-4">
+        <div className="w-full h-full md:w-1/2 ">
+          <Image
+            src={AboutImage}
+            alt="Content creator using StreamFi"
+            className=" shadow-lg h-full"
+            height={300}
+          />
+        </div>
+
+        <div className="w-full md:w-1/2 font-normal text-center h-fit md:text-left">
+          <p className="mb-4">
+            StreamFi is a Web3-powered streaming platform built to give content
+            creators and gamers full control over their earnings. Unlike
+            traditional platforms that take large cuts and delay payouts,
+            StreamFi uses blockchain technology to enable instant crypto
+            tipping, NFT-based memberships, and DeFi staking rewards—all without
+            middlemen.
+          </p>
+          <p>
+            Our mission is to redefine content monetization by offering a
+            transparent, decentralized, and community-driven ecosystem. With
+            smart contracts ensuring secure transactions and DAO governance
+            giving users a say in platform decisions, StreamFi is creating a
+            future where creators earn more, engage directly with their
+            audience, and thrive without restrictions.
+          </p>
+        </div>
+      </div>
+    </Section>
+  );
+}
