+++ data/faqs.ts
@@ -1,0 +1,27 @@
+import { TabItem } from "@/types";
+export const tabs: TabItem[] = [
+    {
+      id: "tab1",
+      title: "What is StreamFi?",
+      content:
+        "StreamFi is a decentralized live-streaming platform that allows creators to monetize their content without intermediaries. Powered by blockchain technology, it ensures 100% earnings retention, direct fan engagement, and community-driven governance.",
+    },
+    {
+      id: "tab2",
+      title: "How do creators earn on StreamFi?",
+      content:
+        "StreamFi is a decentralized live-streaming platform that allows creators to monetize their content without intermediaries. Powered by blockchain technology, it ensures 100% earnings retention, direct fan engagement, and community-driven governance.",
+    },
+    {
+      id: "tab3",
+      title: "Do I need crypto knowledge to use StreamFi?",
+      content:
+        "StreamFi is a decentralized live-streaming platform that allows creators to monetize their content without intermediaries. Powered by blockchain technology, it ensures 100% earnings retention, direct fan engagement, and community-driven governance.",
+    },
+    {
+      id: "tab4",
+      title: " Is StreamFi free to use?",
+      content:
+        "StreamFi is a decentralized live-streaming platform that allows creators to monetize their content without intermediaries. Powered by blockchain technology, it ensures 100% earnings retention, direct fan engagement, and community-driven governance.",
+    },
+  ];
