+++ public/icons/index.ts
@@ -1,0 +1,6 @@
+import CommunityIcon from './community.svg'
+import MoneyLink from './noun-link.svg'
+
+export {
+    CommunityIcon,MoneyLink
+}
