+++ public/images/index.tsx
@@ -1,2 +1,15 @@
 import AboutImage from "./about-image.png";
-export { AboutImage };
+import StripeLogo from "./stripe-logo.png";
+import YoutubeLogo from "./youtube-logo.png";
+import CoinbaseLogo from "./coinbase-logo.png";
+import StreamHero from "./Stream-Fi.png";
+import Discord from "./discord_.png";
+
+export {
+  AboutImage,
+  CoinbaseLogo,
+  StripeLogo,
+  YoutubeLogo,
+  StreamHero,
+  Discord,
+};
