+++ types/index.ts
@@ -1,0 +1,5 @@
+export interface TabItem {
+  id: string;
+  title: string;
+  content: string;
+}
